Code quality: removed unnecessary unboxing

+++ src/org/cacheonix/impl/cache/distributed/partitioned/BucketOwner.java
@@ -574,7 +574,7 @@ public final class BucketOwner implements Wireable {
       final int ownedBucketSize = ownedBuckets.size();
       out.writeInt(ownedBucketSize);
       for (final Integer ownedBucket : ownedBuckets) {
-         out.writeShort(ownedBucket.intValue());
+         out.writeShort(ownedBucket);
       }
 
       //
